rename: 패키지 구조 추가 및 변경
