+++ src/main/java/racingcar/domain/Cars.java
@@ -3,7 +3,7 @@ package racingcar.domain;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
-import racingcar.NumberGenerator;
+import racingcar.util.NumberGenerator;
 
 public class Cars {
