+++ src/main/java/racingcar/util/NumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.util;
 
 public interface NumberGenerator {
     public int generate();
