+++ src/main/java/racingcar/view/InputView.java
@@ -1,7 +1,6 @@
 package racingcar.view;
 
 import java.util.List;
-import racingcar.InputReader;
 
 public class InputView {
     private static final String CAR_NAMES_INPUT_MESSAGE = "경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).";
