+++ src/test/java/racingcar/InputReaderTest.java
@@ -9,6 +9,7 @@ import java.util.Arrays;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.view.InputReader;
 
 class InputReaderTest {
