+++ src/test/java/racingcar/NumberGeneratorTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.util.RandomNumberGenerator;
 
 class NumberGeneratorTest {
